add redirect logic

+++ src/components/common/DealCard.tsx
@@ -9,6 +9,8 @@ import { DEFAULT_DEAL_IMAGE } from '@/config/constants';
 import EllipsisText from './EllipsisText';
 import { useBreakpoints } from '@/hooks/useBreakpoints';
 import Bookmark from './Bookmark';
+import { useUser } from '@/contexts/User';
+import { useRouter } from 'next/router';
 
 export enum DealCardVariant {
   Base = 'base',
@@ -30,16 +32,22 @@ const DealCard = ({
   ...props
 }: DealCardProps) => {
   const classes = useDealCardStyles();
+  const { user } = useUser();
+  const router = useRouter();
   const { isMobile } = useBreakpoints();
 
   const handleAddBookmark = (value: number) => {
-    if (addBookmark) {
+    if (addBookmark && !!user) {
       addBookmark(value);
+    } else {
+      router.push('/login');
     }
   };
   const handleDeleteBookmark = (value: number) => {
-    if (deleteBookmark) {
+    if (deleteBookmark && !!user) {
       deleteBookmark(value);
+    } else {
+      router.push('/login');
     }
   };
   return variant === DealCardVariant.Base ? (
